@@ -1,3 +1,3 @@
 #include "ExoTestState.h"
 
-ExoTestState::ExoTestState(StateMachine *m, X2Robot *exo, DummyTrajectoryGenerator *tg, const char *name) : State(m, name), robot(exo), trajectoryGenerator(tg){};
+ExoTestState::ExoTestState(X2Robot *exo, DummyTrajectoryGenerator *tg, const char *name) : State(name), robot(exo), trajectoryGenerator(tg){};
