@@ -23,7 +23,7 @@ class Standing : public ExoTestState {
     void entry(void);
     void during(void);
     void exit(void);
-    Standing(StateMachine *m, X2Robot *exo, DummyTrajectoryGenerator *tg, const char *name = NULL) : ExoTestState(m, exo, tg, name){};
+    Standing(X2Robot *exo, DummyTrajectoryGenerator *tg, const char *name = "ExoTest Standing State") : ExoTestState(exo, tg, name){};
 };
 
 #endif
