@@ -1,5 +1,5 @@
 /**
- * /file M3DemoState.h
+ * \file M3DemoState.h
  * \author Vincent Crocher
  * \version 0.3
  * \date 2020-07-27
@@ -11,33 +11,21 @@
 #ifndef M3DemoSTATE_H_DEF
 #define M3DemoSTATE_H_DEF
 
-#include <time.h>
-#include <iostream>
-
-#include "RobotM3.h"
 #include "State.h"
+#include "RobotM3.h"
 
-using namespace std;
 
-/**
- * \brief Conversion from a timespec structure to seconds (double)
- *
- */
-double timeval_to_sec(struct timespec *ts);
+class M3DemoMachine;
 
 /**
- * \brief Generic state type for used with M3DemoMachine, providing running time and iterations number.
+ * \brief Generic state type for used with M3DemoMachine, providing running time and iterations number: been superseeded by default state.
  *
  */
 class M3TimedState : public State {
    protected:
-    /**
-    *  \todo Might be good to make these Const
-    *
-    */
-    RobotM3 *robot;                               /*<!Pointer to state machines robot object*/
+    RobotM3 * robot;                               //!< Pointer to state machines robot object
 
-    M3TimedState(StateMachine *m, RobotM3 *M3, const char *name = NULL): State(m, name), robot(M3){};
+    M3TimedState(RobotM3* M3, const char *name = NULL): State(name), robot(M3){spdlog::debug("Created M3TimedState {}", name);};
    private:
     void entry(void) final {
         std::cout
@@ -46,28 +34,10 @@ class M3TimedState : public State {
         << "----------------------------------" << std::endl
         << std::endl;
 
-        //Timing
-        clock_gettime(CLOCK_MONOTONIC, &initTime);
-        lastTime = timeval_to_sec(&initTime);
-
-        elapsedTime=0;
-        iterations=0;
-
         //Actual state entry
         entryCode();
     };
     void during(void) final {
-        //Compute some basic time values
-        struct timespec ts;
-        clock_gettime(CLOCK_MONOTONIC, &ts);
-
-        double now = timeval_to_sec(&ts);
-        elapsedTime = (now-timeval_to_sec(&initTime));
-        dt = now - lastTime;
-        lastTime = now;
-
-        iterations++;
-
         //Actual state during
         duringCode();
     };
@@ -84,27 +54,20 @@ class M3TimedState : public State {
     virtual void entryCode(){};
     virtual void duringCode(){};
     virtual void exitCode(){};
-
-
-   protected:
-    struct timespec initTime;   /*<! Time of state init */
-    double lastTime;            /*<! Time of last during() call (in seconds since state init())*/
-    double elapsedTime;         /*<! Time since state init() in seconds*/
-    double dt;                  /*<! Time between last two during() calls (in seconds)*/
-    unsigned long int iterations;
 };
 
 
 class M3DemoState : public M3TimedState {
 
    public:
-    M3DemoState(StateMachine *m, RobotM3 *M3, const char *name = "M3 Test State"):M3TimedState(m, M3, name){};
+    M3DemoState(RobotM3 * M3, const char *name = "M3 Test State"):M3TimedState(M3, name){};
 
     void entryCode(void);
     void duringCode(void);
     void exitCode(void);
 
     VM3 qi, Xi;
+    double f=0.0;
 };
 
 
@@ -116,7 +79,7 @@ class M3DemoState : public M3TimedState {
 class M3CalibState : public M3TimedState {
 
    public:
-    M3CalibState(StateMachine *m, RobotM3 *M3, const char *name = "M3 Calib State"):M3TimedState(m, M3, name){};
+    M3CalibState(RobotM3 * M3, const char *name = "M3 Calib State"):M3TimedState(M3, name){};
 
     void entryCode(void);
     void duringCode(void);
@@ -139,7 +102,7 @@ class M3CalibState : public M3TimedState {
 class M3MassCompensation : public M3TimedState {
 
    public:
-    M3MassCompensation(StateMachine *m, RobotM3 *M3, const char *name = "M3 Mass Compensation"):M3TimedState(m, M3, name){};
+    M3MassCompensation(RobotM3 * M3, const char *name = "M3 Mass Compensation"):M3TimedState(M3, name){};
 
     void entryCode(void);
     void duringCode(void);
@@ -158,7 +121,7 @@ class M3MassCompensation : public M3TimedState {
 class M3EndEffDemo : public M3TimedState {
 
    public:
-    M3EndEffDemo(StateMachine *m, RobotM3 *M3, const char *name = "M3 Velocity Control Demo"):M3TimedState(m, M3, name){};
+    M3EndEffDemo(RobotM3 * M3, const char *name = "M3 Velocity Control Demo"):M3TimedState(M3, name){};
 
     void entryCode(void);
     void duringCode(void);
@@ -172,7 +135,7 @@ class M3EndEffDemo : public M3TimedState {
 class M3DemoImpedanceState : public M3TimedState {
 
    public:
-    M3DemoImpedanceState(StateMachine *m, RobotM3 *M3, const char *name = "M3 Demo Impedance State"):M3TimedState(m, M3, name){};
+    M3DemoImpedanceState(RobotM3 * M3, const char *name = "M3 Demo Impedance State"):M3TimedState(M3, name){};
 
     void entryCode(void);
     void duringCode(void);
@@ -180,7 +143,7 @@ class M3DemoImpedanceState : public M3TimedState {
 
    private:
     VM3 Xi;
-    double k = 700;     //! Impedance proportional gain (spring)
+    double k = 2000;     //! Impedance proportional gain (spring)
     double d = 2;       //! Impedance derivative gain (damper)
     bool init=false;
 
@@ -192,13 +155,33 @@ class M3DemoImpedanceState : public M3TimedState {
 
 
 /**
+ * \brief Teleoperation state
+ *
+ */
+class M3TeleopState : public M3TimedState {
+
+   public:
+    M3TeleopState(RobotM3 * M3, M3DemoMachine &sm, const char *name = "M3 Teleop State"):M3TimedState(M3, name), SM(sm) {};
+
+    void entryCode(void);
+    void duringCode(void);
+    void exitCode(void);
+
+   private:
+    M3DemoMachine &SM;
+    VM3 F;
+    VM3 Xi;
+};
+
+
+/**
  * \brief Path contraint with viscous assistance.
  *
  */
 class M3DemoPathState : public M3TimedState {
 
    public:
-    M3DemoPathState(StateMachine *m, RobotM3 *M3, const char *name = "M3 Demo Path State"):M3TimedState(m, M3, name){};
+    M3DemoPathState(RobotM3 * M3, const char *name = "M3 Demo Path State"):M3TimedState(M3, name){};
 
     void entryCode(void);
     void duringCode(void);
@@ -216,13 +199,13 @@ class M3DemoPathState : public M3TimedState {
 
 
 /**
- * \brief Point to tpoint position control with min jerk trajectory interpolation
+ * \brief Point to point position control with min jerk trajectory interpolation
  *
  */
 class M3DemoMinJerkPosition: public M3TimedState {
 
    public:
-    M3DemoMinJerkPosition(StateMachine *m, RobotM3 *M3, const char *name = "M3 Demo Minimum Jerk Position"):M3TimedState(m, M3, name){};
+    M3DemoMinJerkPosition(RobotM3 * M3, const char *name = "M3 Demo Minimum Jerk Position"):M3TimedState(M3, name){};
 
     void entryCode(void);
     void duringCode(void);
@@ -232,8 +215,8 @@ class M3DemoMinJerkPosition: public M3TimedState {
     static const unsigned int TrajNbPts=4;
     unsigned int TrajPtIdx=0;
     double startTime;
-    VM3 TrajPt[TrajNbPts]={VM3(-0.55, 0, 0), VM3(-0.7, 0, -0.2), VM3(-0.7, 0.1, -0.1), VM3(-0.7, -0.1, -0.1)};
-    double TrajTime[TrajNbPts]={5, 3, 0.5, 0.5};
+    VM3 TrajPt[TrajNbPts]={VM3(-0.6, 0, 0), VM3(-0.6, 0., 0.2), VM3(-0.6, 0., -0.2), VM3(-0.6, 0, 0.2)};
+    double TrajTime[TrajNbPts]={5, 5, 5, 1};
     VM3 Xi, Xf;
     double T;
     float k_i=1.; //Integral gain
@@ -248,7 +231,7 @@ class M3DemoMinJerkPosition: public M3TimedState {
 class M3SamplingEstimationState : public M3TimedState {
 
    public:
-    M3SamplingEstimationState(StateMachine *m, RobotM3 *M3, const char *name = "M3 Sampling time estimation State"):M3TimedState(m, M3, name){};
+    M3SamplingEstimationState(RobotM3 * M3, const char *name = "M3 Sampling time estimation State"):M3TimedState(M3, name){};
 
     void entryCode(void);
     void duringCode(void);
