@@ -1,6 +1,6 @@
 #include "X2DemoMachineROS.h"
 
-X2DemoMachineROS::X2DemoMachineROS(X2Robot *robot, X2DemoState *x2DemoState, ros::NodeHandle& nodeHandle):
+X2DemoMachineROS::X2DemoMachineROS(X2Robot *robot, std::shared_ptr<X2DemoState> x2DemoState, ros::NodeHandle& nodeHandle):
         robot_(robot),
         x2DemoState_(x2DemoState),
         nodeHandle_(&nodeHandle)
